Add files via upload
Here is the latest code for the RFID Smartlocks. It allows for opening the lock using an RFID key but also allows access via the private network (WiFi). This means you can open the lock by sending code using something like a main control station or even change the UID code used to open the lock.